DynamoDB features behind feature flag in auth_lib
