+++ serverless-aws-sam/src/auth/auth_lib/Cargo.toml
@@ -3,12 +3,15 @@ name = "auth_lib"
 version = "0.1.0"
 edition = "2024"
 
+[features]
+default = []
+dynamodb = ["aws-sdk-dynamodb", "argon2"]
+
 [dependencies]
 lambda_runtime = "0.14.3"
 serde = "1.0.219"
 serde_json = "1"
 rand = "0.8"
-argon2 = "0.5"
 uuid = { version = "1.18", features = ["v4", "rng"] }
 jsonwebtoken = "9.3.1"
 tokio = { version = "1", features = ["macros"] }
@@ -18,6 +21,9 @@ tracing-subscriber = { version = "0.3", default-features = false, features = [
 ] }
 aws-config = "1.8.5"
 aws-credential-types = "1.2.5"
-aws-sdk-dynamodb = "1.87.0"
 aws-sdk-ssm = "1.90.0"
 chrono = "0.4.41"
+
+# Optional and tied to the "dynamodb" feature
+aws-sdk-dynamodb = { version = "1.87.0", optional = true }
+argon2 = { version = "0.5", optional = true }
