+++ serverless-aws-sam/src/auth/auth_lib/src/lib.rs
@@ -1,14 +1,13 @@
-use argon2::{
-    Argon2,
-    password_hash::{PasswordHash, PasswordHasher, PasswordVerifier, SaltString, rand_core::OsRng},
-};
 use lambda_runtime::Error;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use std::collections::HashMap;
 
+#[cfg(feature = "dynamodb")]
 pub mod models;
+#[cfg(feature = "dynamodb")]
+pub use crate::models::PasswordManager;
 pub mod tokens;
 mod validate_body;
 pub use crate::validate_body::ValidateBodyEnum;
@@ -60,33 +59,6 @@ pub enum LambdaReturnEnum<T> {
     Response(Result<Response, Error>),
 }
 
-pub struct PasswordManager;
-
-impl PasswordManager {
-    /// Hash a password with a random salt using Argon2
-    pub fn hash_password(password: &str) -> Result<String, argon2::password_hash::Error> {
-        let salt = SaltString::generate(&mut OsRng);
-        let argon2 = Argon2::default();
-        let password_hash = argon2.hash_password(password.as_bytes(), &salt)?;
-        Ok(password_hash.to_string())
-    }
-
-    /// Verify a password against a stored hash
-    pub fn verify_password(
-        password: &str,
-        hash: &str,
-    ) -> Result<bool, argon2::password_hash::Error> {
-        let parsed_hash = PasswordHash::new(hash)?;
-        let argon2 = Argon2::default();
-
-        match argon2.verify_password(password.as_bytes(), &parsed_hash) {
-            Ok(()) => Ok(true),
-            Err(argon2::password_hash::Error::Password) => Ok(false),
-            Err(e) => Err(e),
-        }
-    }
-}
-
 const COLORS: [&'static str; 7] = ["blue", "green", "orange", "yellow", "sky", "purple", "pink"];
 pub fn get_random_color() -> &'static str {
     let mut rng = rand::thread_rng();
@@ -116,24 +88,5 @@ mod tests {
     fn get_random_color_test() {
         let color = get_random_color();
         assert!(COLORS.contains(&color), "Did not get a color")
-    }
-
-    #[test]
-    fn password_hashing_and_verification_test() {
-        let password = "my_secure_password123";
-
-        // Hash the password
-        let hash = PasswordManager::hash_password(password).unwrap();
-        println!("Generated hash: {}", hash);
-
-        // Verify correct password
-        let is_valid = PasswordManager::verify_password(password, &hash).unwrap();
-        assert!(is_valid);
-
-        // Verify incorrect password
-        let is_invalid = PasswordManager::verify_password("wrong_password", &hash).unwrap();
-        assert!(!is_invalid);
-
-        println!("Password verification test passed!");
     }
 }
